refactor: 코드 구조 MVC 패턴으로 개선 - Input클래스 Output클래스 분리 - model에서는 CarManager,GameManaget로 분리 - GameContoller 분리 - Valdation 분리
