+++ README.md
@@ -34,25 +34,25 @@
 - [x] 자세한 과제 진행 방법은 프리코스 진행 가이드 문서를 참고한다.
 
 ### 기능 요구사항
-- [ ] 주어진 횟수 동안 n대의 자동차는 전진 또는 멈출 수 있다.
-- [ ] 각 자동차에 이름을 부여할 수 있다. 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
-- [ ] 자동차 이름은 쉼표(,)를 기준으로 구분하며 이름은 5자 이하만 가능하다.
-- [ ] 사용자는 몇 번의 이동을 할 것인지를 입력할 수 있어야 한다.
+- [x] 주어진 횟수 동안 n대의 자동차는 전진 또는 멈출 수 있다.
+- [x] 각 자동차에 이름을 부여할 수 있다. 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
+- [x] 자동차 이름은 쉼표(,)를 기준으로 구분하며 이름은 5자 이하만 가능하다.
+- [x] 사용자는 몇 번의 이동을 할 것인지를 입력할 수 있어야 한다.
 - [ ] 전진하는 조건은 0에서 9 사이에서 무작위 값을 구한 후 무작위 값이 4 이상일 경우이다.
 - [ ] 자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한 명 이상일 수 있다.
 - [ ] 우승자가 여러 명일 경우 쉼표(,)를 이용하여 구분한다.
 - [ ] 사용자가 잘못된 값을 입력할 경우 "[ERROR]"로 시작하는 메시지와 함께 Error를 발생시킨 후 애플리케이션은 종료되어야 한다.
 
 ### 입출력 요구 사항
-- [ ] 주어진대로 구현하였다.
+- [x] 주어진대로 구현하였다.
 
 ### 프로그래밍 요구사항
-- [ ] Node.js 20.17.0 버전에서 실행 가능해야 한다.
-- [ ] 프로그램 실행의 시작점은 App.js의 run()이다.
-- [ ] package.json 파일은 변경할 수 없으며, 제공된 라이브러리와 스타일 라이브러리 이외의 외부 라이브러리는 사용하지 않는다.
-- [ ] 프로그램 종료 시 process.exit()를 호출하지 않는다.
-- [ ] 프로그래밍 요구 사항에서 달리 명시하지 않는 한 파일, 패키지 등의 이름을 바꾸거나 이동하지 않는다.
-- [ ] 자바스크립트 코드 컨벤션을 지키면서 프로그래밍한다.
+- [x] Node.js 20.17.0 버전에서 실행 가능해야 한다.
+- [x] 프로그램 실행의 시작점은 App.js의 run()이다.
+- [x] package.json 파일은 변경할 수 없으며, 제공된 라이브러리와 스타일 라이브러리 이외의 외부 라이브러리는 사용하지 않는다.
+- [x] 프로그램 종료 시 process.exit()를 호출하지 않는다.
+- [x] 프로그래밍 요구 사항에서 달리 명시하지 않는 한 파일, 패키지 등의 이름을 바꾸거나 이동하지 않는다.
+- [x] 자바스크립트 코드 컨벤션을 지키면서 프로그래밍한다.
 
 ### 프로그래밍 요구사항2
 - [ ] indent(인덴트, 들여쓰기) depth를 3이 넘지 않도록 구현한다. 2까지만 허용한다.
@@ -63,15 +63,15 @@
 
 # 🤔 개발하면서 신경쓴 부분
 ### 공통피드백
-- [ ] 출력함수대신 디버거를 사용
-- [ ] js에서 제공하는 API를 적극 활용하기
+- [x] 출력함수대신 디버거를 사용
+- [x] js에서 제공하는 API를 적극 활용하기
 - [ ] 이름을 통해 의도를 나타내고, 이름 축약하지 않기
 - [ ] 요구사항 완전히 준수하기
 - [x] 코드 포매팅 활용하기
-- [ ] 일단 굴러가게 만든 뒤, 리팩토링 하기
+- [x] 일단 굴러가게 만든 뒤, 리팩토링 하기
 
 ### 피어리뷰
-- [ ] private 필요할 경우 사용하기
+- [x] private 필요할 경우 사용하기
 - [ ] 에러처리 상수처리하기
 - [ ] 예외처리에 신경 많이써서 많은 예외에 대해 대비하기
 - [ ] Object.freeze를 사용해서 상수를 불변하게 관리하기
